requirements.txt
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
google-auth==2.9.1
googleapis-common-protos==1.56.4
Flask-SQLAlchemy==2.5.1
Flask==2.0.3
pandas==1.4.2